v0.3.8 Restructured the event module.  - Renamed event folder to event_types  - Split "event.hpp" into "event_base.hpp" and "events.hpp"   * "event_base.hpp" acts as a base for all event types.   * "events.hpp" acts as a common include file for the event module. - Added events into the precompiled header.
