+++ tractor/include/tractor.hpp
@@ -20,19 +20,7 @@
 #include "tractor/logger.hpp"
 #include "tractor/window.hpp"
 
-#include "tractor/events/event.hpp"
-#include "tractor/events/event_application.hpp"
-#include "tractor/events/event_audio.hpp"
-#include "tractor/events/event_controller.hpp"
-#include "tractor/events/event_display.hpp"
-#include "tractor/events/event_gesture.hpp"
-#include "tractor/events/event_joystick.hpp"
-#include "tractor/events/event_keyboard.hpp"
-#include "tractor/events/event_mouse.hpp"
-#include "tractor/events/event_render.hpp"
-#include "tractor/events/event_system.hpp"
-#include "tractor/events/event_touch.hpp"
-#include "tractor/events/event_window.hpp"
+#include "tractor/events.hpp"
 
 namespace trac
 {
