+++ tractor/include/tractor/event_types/event_audio.hpp
@@ -10,7 +10,7 @@
 #define EVENT_AUDIO_HPP_
 
 // Related header include
-#include "event.hpp"
+#include "event_base.hpp"
 
 // External libraries header includes
 #include <SDL_audio.h>
