+++ tractor/src/event_types/event_joystick.cpp
@@ -11,7 +11,7 @@
 #include "tractor_pch.hpp"
 
 // Related header include
-#include "events/event_joystick.hpp"
+#include "event_types/event_joystick.hpp"
 
 // External libraries header includes
 #include <SDL_joystick.h>
