+++ tractor/src/event_types/event_render.cpp
@@ -10,7 +10,7 @@
 #include "tractor_pch.hpp"
 
 // Related header include
-#include "events/event_render.hpp"
+#include "event_types/event_render.hpp"
 
 // External libraries header includes
 #include <SDL_timer.h>
