+++ tractor/src/events.cpp
@@ -1,17 +1,16 @@
 /**
- * @file	event.cpp
- * @brief	Main source file for the event module. See event.hpp for more information.
+ * @file	events.cpp
+ * @brief	Main source file for the event module. See events.hpp for more information.
  * 
  * @author	Erlend Elias Isachsen
- * @date	26.07.2023
- * 
-*/
+ * @date	27.08.2023
+ */
 
 // Precompiled header include
 #include "tractor_pch.hpp"
 
 // Related header include
-#include "events/event.hpp"
+#include "events.hpp"
 
 // Project includes
 #include "logger.hpp"
@@ -233,53 +232,6 @@ namespace trac
 	}
 
 	/**
-	 * @brief	Inserter operator for the Event class. This function is used to insert an event into an output stream (i.e. to print event data).
-	 * 
-	 * @param os	The output stream to insert the event into.
-	 * @param e	The event to insert into the output stream.
-	 * @return std::ostream&	The output stream with the event inserted.
-	 */
-	std::ostream& operator<<(std::ostream& os, const Event& e)
-	{
-		return os << e.ToString();
-	}
-
-	/**
-	 * @brief	Inserter operator for pointers to the Event class. This function is used to insert an event into an output stream (i.e. to print event data).
-	 * 
-	 * @param os	The output stream to insert the event into.
-	 * @param e	The event to insert into the output stream.
-	 * @return std::ostream&	The output stream with the event inserted.
-	 */
-	std::ostream& operator<<(std::ostream& os, const std::shared_ptr<Event> e)
-	{
-		return os << e->ToString();
-	}
-
-	/**
-	 * @brief	Converts the event to a string representation. This function is used to print the event data to the console.
-	 * 
-	 * @return std::string	The string representation of the event.
-	 */
-	std::string Event::ToString() const
-	{
-		return GetName();
-	}
-
-	/**
-	 * @brief	Returns whether the event is in the provided category or not.
-	 * 
-	 * @param category	The category to check.
-	 * @return bool	True if the event is in the provided category, false otherwise.
-	 * @retval True	The event is in the provided category.
-	 * @retval False	The event is not in the provided category.
-	 */
-	bool Event::IsInCategory(EventCategory category) const
-	{
-		return GetCategoryFlags() & category;
-	}
-
-	/**
 	 * @brief	Returns the event type for the for a given event reference.
 	 * 
 	 * @param e	The event to get the type for.
@@ -334,4 +286,5 @@ namespace trac
 	{
 		return engine_queue_s_;
 	}
-} // Namespace trac
+
+} // namespace trac
